Module trivial_quest::plugins::physics

Author: Ethan

Physics are an essential part of any video game. This plugin handles basic mechanics, including movement, gravity, and collisions.

There are two main components: Movable and Fixed, both of which are managed by the physics engine. An object cannot be both Movable and Fixed. For example, players and enemies should be Movable, while platforms and walls should be Fixed.

When managing the position of a Movable object, please use Movable::pos rather than setting the transform directly (except initally). The physics engine will automatically handle updating the transform.

Additional components control behavior and interaction with other physics objects. They include:

Gravity makes a Movable affected by gravity.

Collider makes a Movable collide with Fixed objects that are also Collider.

During Bevy’s Update, the physics engine will apply gravity, check for collisions, and move all Movables by updating their transforms.

A good thing to do now would be to read the instructions for how to use Movable.

Structs

Fixed

Gravity

Use this component on a Movable to enable gravity.

Movable

A movable physics object. For a sprite to have its movement managed by the physics engine (i.e., to support collisions and gravity), you must do two things:

PhysicsPlugin

Enums

Collider

Constants

GRAVITY

The acceleration to due gravity.