Module trivial_quest::plugins::physics::movement

Movement systems.

Structs

CollisionEvent

A CollisionEvent { movable, fixed, timestamp } is produced when a solid movable collides with a solid fixed.

Functions

carry_out_movement

Realizes the now-satisfiable requested movement of movables.

move_and_collide

Collides solid Movable objects (with position and velocity) with Fixed objects, preventing movement of solid movables through solid fixed objects (see ColliderBehavior::Solid).