Module trivial_quest::plugins::render

Author: Sidharth Rao

The render plugin handles game rendering, texturing, and cameras.

It creates a MainCamera entity and follows an entity with the CameraFollow component (usually the player). If no entity has the CameraFollow component, the camera will remain stationary. There can be at most one entity with the MainCamera component.

The plugin also loads sprites by querying for entities with the [AssetPath] component and the [Transform] component. It then replaces the [Transform] component with a [SpriteBundle] component containing the texture loaded from the [AssetPath] component and the original [Transform] component.

Schedule: - Startup: Sets up the main camera. - Update: Loads sprites. Only processes entities with the [AssetPath] component that has changed. This means that users can change the texture of an entity by changing the [AssetPath] component. - PostUpdate: Follows the entity (or no entity) with the CameraFollow component.

Structs

Asset

Contains the path to an asset and its size.

CameraFollow

If exactly one entity has this component, the camera will follow it.

CameraZoom

Sets how much to zoom the camera. The default zoom is 1.0. A zoom of 2.0 will make objects appear twice as small.

MainCamera

Main camera.

RenderPlugin

Handles game rendering, texturing, and cameras.

Functions

zoom_camera

Zooms the MainCamera based on the CameraZoom resource.